fix: move data class to data package
