+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/data/GenericUserResponse.kt
@@ -1,0 +1,12 @@
+package io.tolgee.ee.data
+
+@Suppress("PropertyName")
+class GenericUserResponse {
+  var sub: String? = null
+  var name: String? = null
+
+  var given_name: String? = null
+
+  var family_name: String? = null
+  var email: String? = null
+}
