+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/data/OAuth2TokenResponse.kt
@@ -1,0 +1,7 @@
+package io.tolgee.ee.data
+
+@Suppress("PropertyName")
+class OAuth2TokenResponse(
+  val id_token: String,
+  val scope: String,
+)
